\name{blupADC-package}
\alias{blupADC-package}
\alias{blupADC}
\docType{package}
\title{
\packageTitle{blupADC}
}
\description{
\packageDescription{blupADC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{blupADC}
\packageIndices{blupADC}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{blupADC}

Maintainer: \packageMaintainer{blupADC}
}
\references{
~~ Literature or other references for background information ~~
}