\name{blupADC-package}
\alias{blupADC-package}
\alias{blupADC}
\docType{package}
\title{
\packageTitle{blupADC}
}
\description{
\packageDescription{blupADC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{blupADC}
\packageIndices{blupADC}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{blupADC}

Maintainer: \packageMaintainer{blupADC}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
